backend/requirements.txt
flask
nltk
pandas
openpyxl
python-docx
PyPDF2
pdfminer.six
flask-cors
requests
